third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-into-inlined-scripts.html

<html>
<head>

<script>
function f1()
{
    return 1; // Breakpoint.
}
f1();
</script>

<script>
function f2()
{
    return 2;
}
f2();
</script>

<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/debugger-test.js"></script>

<script>

function test() {
  var numberOfStepInto = 6;

  SourcesTestRunner.startDebuggerTest(step1, true);

  function step1() {
    SourcesTestRunner.showScriptSource('debugger-step-into-inlined-scripts.html', step2);
  }

  function step2(sourceFrame) {
    TestRunner.addResult('Script source was shown.');
    SourcesTestRunner.setBreakpoint(sourceFrame, 6, '', true);
    SourcesTestRunner.waitUntilPaused(step3);
    TestRunner.reloadPage(completeTest);
  }

  function step3() {
    var actions = ['Print'];  // First pause on breakpoint.
    for (var i = 0; i < numberOfStepInto; ++i)
      actions.push('StepInto', 'Print');
    actions.push('Resume');
    SourcesTestRunner.waitUntilPausedAndPerformSteppingActions(actions);
  }

  function completeTest() {
    SourcesTestRunner.completeDebuggerTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that debugger StepInto will step through inlined scripts.
</p>
</body>
</html>
