third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-step/debugger-step-into-inlined-scripts-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that debugger StepInto will step through inlined scripts.

Script source was shown.
Call stack:
    0) f1 (debugger-step-into-inlined-scripts.html:7)
    1)  (debugger-step-into-inlined-scripts.html:9)

Executing StepInto...
Call stack:
    0) f1 (debugger-step-into-inlined-scripts.html:7)
    1)  (debugger-step-into-inlined-scripts.html:9)

Executing StepInto...
Call stack:
    0)  (debugger-step-into-inlined-scripts.html:10)

Executing StepInto...
Call stack:
    0)  (debugger-step-into-inlined-scripts.html:17)

Executing StepInto...
Call stack:
    0) f2 (debugger-step-into-inlined-scripts.html:15)
    1)  (debugger-step-into-inlined-scripts.html:17)

Executing StepInto...
Call stack:
    0) f2 (debugger-step-into-inlined-scripts.html:15)
    1)  (debugger-step-into-inlined-scripts.html:17)

Executing StepInto...
Call stack:
    0)  (debugger-step-into-inlined-scripts.html:18)

Executing Resume...
Page reloaded.
